+++ app/src/main/java/app/hillforts/views/hillfortList/HillfortListPresenter.kt
@@ -3,10 +3,13 @@ package app.hillforts.views.hillfortList
 import android.content.Intent
 import app.hillforts.main.MainApp
 import app.hillforts.models.HillfortModel
+import app.hillforts.models.UserModel
 import app.hillforts.views.editLocation.HillfortMapsView
 import app.hillforts.views.hillfort.HillfortView
 import app.hillforts.views.login.LoginView
 import app.hillforts.views.settings.SettingsView
+import com.google.firebase.auth.FirebaseAuth
+import org.jetbrains.anko.doAsync
 import org.jetbrains.anko.intentFor
 import org.jetbrains.anko.startActivity
 import org.jetbrains.anko.startActivityForResult
@@ -20,7 +23,7 @@ class HillfortListPresenter(val view: HillfortListView) {
         app = view.application as MainApp
     }
 
-    fun doGetHillforts() = app.unified.findAllHillfortsForUser(app.appUser)
+    fun doGetHillforts() = app.unified.findAllHillforts()
 
     fun doAddHillfort() {
         view.startActivityForResult<HillfortView>(0)
@@ -41,6 +44,9 @@ class HillfortListPresenter(val view: HillfortListView) {
     }
 
     fun doLogout() {
-        view.startActivityForResult<LoginView>(0)
+        view.finish()
+        FirebaseAuth.getInstance().signOut()
+        app.unified.clear()
+        view.startActivity<LoginView>()
     }
 }