+++ app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.14.0 - 05-02-2020 ***(Merged)***
+  - TASK - [MOB-DEV-15]  - Lab 11, Firebase database completed.
+                           Landscape views added.
+
+## 1.13.0 - 04-02-2020 ***(Merged)***
+  - TASK - [MOB-DEV-14]  - Lab 10, Firebase authentication completed.
+
 ## 1.12.1 - 04-02-2020 ***(Merged)***
   - TASK - [HOTFIX-1]  - Hotfix for hillfort view layout, misalignment in fields.